% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply a filtering criterion}
\usage{
apply_filter(d, var, values)
}
\arguments{
\item{d}{input data frame.}

\item{var}{variable name (as a quoted character string).}

\item{values}{vector of values used as filtering criterion: for qualitative variables, a vector of valid values; for quantitative variables, the valid range.}
}
\value{
A data frame, comprising only the subset of rows of the input data frame that match the filter.
}
\description{
Apply a filtering criterion
}
\examples{
summary(ctd)
summary(apply_filter(ctd, "month", c("Mar", "Apr", "May")))
hist(ctd$sal, breaks=100)
ctd_clean <- apply_filter(ctd, "sal", c(37.7, 38.4))
hist(ctd_clean$sal, breaks=100)
}
