% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_range.R
\name{pretty_range}
\alias{pretty_range}
\title{Compute a nice looking range for a slider}
\usage{
pretty_range(x)
}
\arguments{
\item{x}{an object coercible to numeric by as.numeric.}
}
\value{
A vector with two elements: the nice looking range
}
\description{
Compute a nice looking range for a slider
}
\examples{
range(ctd$temp)
pretty_range(ctd$temp)
range(ctd$date)
pretty_range(ctd$date)
}
